% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tools.R
\name{freegeoip}
\alias{freegeoip}
\title{freegeoip
Locate IP address using freegeoip.net service}
\usage{
freegeoip(ip, format = ifelse(length(ip) == 1, "list", "dataframe"))
}
\arguments{
\item{ip}{}

\item{format}{}
}
\description{
freegeoip
Locate IP address using freegeoip.net service
}

