% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tools.R
\name{getProtocolFromPort}
\alias{getProtocolFromPort}
\title{getProtocolFromPort}
\usage{
getProtocolFromPort(port, istcp)
}
\arguments{
\item{port}{well known port number}

\item{istcp}{boolean, True for TCP, False for UDP
212}
}
\description{
given a well known port number and true for TCP, False for UDP returns the protocol name that runs behind
}

