% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tools.R
\name{hex2ip}
\alias{hex2ip}
\title{hex2ip
Transform an 8 bytes hexadecimal string to ip address}
\usage{
hex2ip(hex)
}
\arguments{
\item{hex}{ip <- hex2ip("c0a80001")}
}
\description{
hex2ip
Transform an 8 bytes hexadecimal string to ip address
}

