% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tools.R
\name{ip2hex}
\alias{ip2hex}
\title{ip2hex
Transform an ip address to hexadecimal}
\usage{
ip2hex(ip)
}
\arguments{
\item{ip}{}
}
\description{
ip2hex
Transform an ip address to hexadecimal
}
\examples{
hex <- ip2hex("192.168.0.1")
}

