% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tools.R
\name{ip2long}
\alias{ip2long}
\title{ip2long
Transforma una IP "192.168.0.1" a integer 3232235521}
\usage{
ip2long(ip)
}
\arguments{
\item{ip}{}
}
\description{
ip2long
Transforma una IP "192.168.0.1" a integer 3232235521
}
\examples{
ip <- ip2long("192.168.0.1")
}

