% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tools.R
\name{ip_in_CIDR}
\alias{ip_in_CIDR}
\title{ip_in_CIDR
Check if IP address (string) is in a CIDR range (string)}
\usage{
ip_in_CIDR(ip, cidr)
}
\arguments{
\item{ip}{}

\item{cidr}{}
}
\description{
ip_in_CIDR
Check if IP address (string) is in a CIDR range (string)
}

