% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tools.R
\name{long2ip}
\alias{long2ip}
\title{long2ip Convert integer IP address 3232235521 to character "192.168.0.1"}
\usage{
long2ip(longip)
}
\arguments{
\item{longip}{}
}
\description{
long2ip Convert integer IP address 3232235521 to character "192.168.0.1"
}

